+++ src/interrogate/cli.py
@@ -103,6 +103,14 @@ from interrogate import utils
     help="Ignore nested functions and methods.",
 )
 @click.option(
+    "-C",
+    "--ignore-nested-classes",
+    is_flag=True,
+    default=False,
+    show_default=True,
+    help="Ignore nested classes.",
+)
+@click.option(
     "-p",
     "--ignore-private",
     is_flag=True,
@@ -122,6 +130,14 @@ from interrogate import utils
     default=False,
     show_default=True,
     help="Ignore methods with property setter/getter decorators.",
+)
+@click.option(
+    "-S",
+    "--ignore-setters",
+    is_flag=True,
+    default=False,
+    show_default=True,
+    help="Ignore methods with property setter decorators.",
 )
 @click.option(
     "-s",
@@ -190,6 +206,18 @@ from interrogate import utils
         "change from an existing badge of the same path."
     ),
 )
+@click.option(
+    "--badge-format",
+    type=click.Choice(["svg", "png"], case_sensitive=False),
+    default=None,
+    show_default="svg",
+    help=(
+        "File format for the generated badge. Used with the "
+        "`-g/--generate-badge` flag. [default: svg]"
+        "\n\nNOTE: To generate a PNG file, interrogate must be installed "
+        "with `interrogate[png]`, i.e. `pip install interrogate[png]`."
+    ),
+)
 @click.help_option("-h", "--help")
 @click.argument(
     "paths",
@@ -229,10 +257,20 @@ def main(ctx, paths, **kwargs):
     .. versionadded:: 1.2.0 ``--color``/``--no-color``
     .. versionadded:: 1.3.0 ``--ignore-property-decorators``
     .. versionadded:: 1.3.0 config parsing support for setup.cfg
+    .. versionadded:: 1.4.0 ``--badge-format``
+    .. versionadded:: 1.4.0 ``--ignore-nested-classes``
+    .. versionadded:: 1.4.0 ``--ignore-setters``
 
     .. versionchanged:: 1.3.1 only generate badge if results change from
         an existing badge.
     """
+    gen_badge = kwargs["generate_badge"]
+    if kwargs["badge_format"] is not None and gen_badge is None:
+        raise click.BadParameter(
+            "The `--badge-format` option must be used along with the `-g/"
+            "--generate-badge option."
+        )
+
     if not paths:
         paths = (os.path.abspath(os.getcwd()),)
 
@@ -250,7 +288,9 @@ def main(ctx, paths, **kwargs):
         ignore_magic=kwargs["ignore_magic"],
         ignore_module=kwargs["ignore_module"],
         ignore_nested_functions=kwargs["ignore_nested_functions"],
+        ignore_nested_classes=kwargs["ignore_nested_classes"],
         ignore_property_decorators=kwargs["ignore_property_decorators"],
+        ignore_property_setters=kwargs["ignore_setters"],
         ignore_private=kwargs["ignore_private"],
         ignore_regex=kwargs["ignore_regex"],
         ignore_semiprivate=kwargs["ignore_semiprivate"],
@@ -272,8 +312,11 @@ def main(ctx, paths, **kwargs):
             results, kwargs["output"], kwargs["verbose"]
         )
 
-    if kwargs["generate_badge"] is not None:
-        output_path = badge_gen.create(kwargs["generate_badge"], results)
+    if gen_badge is not None:
+        badge_format = kwargs["badge_format"]
+        output_path = badge_gen.create(
+            gen_badge, results, output_format=badge_format
+        )
         if not is_quiet:
             click.echo("Generated badge to {}".format(output_path))