+++ docs/changelog.rst
@@ -1,8 +1,17 @@
 Changelog
 =========
 
-1.3.3 (UNRELEASED)
+.. short-log
+
+1.4.0 (2021-05-14)
 ------------------
+
+Added
+^^^^^
+
+* Support for generating the status badge as a PNG file with a new ``--badge-format`` flag (`#70 <https://github.com/econchick/interrogate/issues/70>`_).
+* Add new option ``-C`` / ``--ignore-nested-classes`` to ignore – you guessed it – nested classes (`#65 <https://github.com/econchick/interrogate/issues/65>`_).
+* Add new option ``-S`` / ``--ignore-setters`` to ignore property setter decorators (`#68 <https://github.com/econchick/interrogate/issues/68>`_).
 
 .. short-log
 
@@ -26,8 +35,6 @@ Removed
 
 * Support for Python 3.5 – thank you `s-weigand <https://github.com/econchick/interrogate/pull/58>`_!
 
-
-.. short-log
 
 1.3.1 (2020-09-03)
 ------------------